+++ src/common/FinanceNavigation/constants.js
@@ -1,6 +1,5 @@
-// eslint-disable-next-line import/prefer-default-export
 export const FINANCE_NAVIGATION_TABS = {
-  FISCAL_YAER: 'fiscalyear',
+  FISCAL_YEAR: 'fiscalyear',
   LEDGER: 'ledger',
   GROUPS: 'groups',
   FUND: 'fund',