+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fiscal year is empty after adding funds to new group. Refs UIF-345.
 * Tech debt: reducing code smells, usage of consts and resources from stripes-acq. Refs UIF-346.
 * Allow user to base encumbrance on "Initial encumbrance" during rollover. Refs UIF-324.
+* Edit restrictExpenditures, restrictEncumbrance ledger toggles in create/edit form. Refs UIF-349.
 
 ## [3.0.1](https://github.com/folio-org/ui-finance/tree/v3.0.1) (2021-11-02)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v3.0.0...v3.0.1)