+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add a return to Finance default search to app context menu dropdown. Refs UIF-350.
 * Display transactions as Voided for canceled invoices. Refs UIF-316.
 * Set enforce limit toggles to True by default in Rollover settings form. Refs UIF-352.
+* Settings > Finance > change focus. Refs UIF-354.
 
 ## [3.0.1](https://github.com/folio-org/ui-finance/tree/v3.0.1) (2021-11-02)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v3.0.0...v3.0.1)