+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Set enforce limit toggles to True by default in Rollover settings form. Refs UIF-352.
 * Settings > Finance > change focus. Refs UIF-354.
 * Refactor psets away from backend ".all" permissions.Refs UIF-333.
+* Voided amount not displayed for voided transactions. Refs UINV-358.
 
 ## [3.0.1](https://github.com/folio-org/ui-finance/tree/v3.0.1) (2021-11-02)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v3.0.0...v3.0.1)